+++ core/chains/evm/client/erroring_node.go
@@ -130,6 +130,10 @@ func (e *erroringNode) StateAndLatest() (NodeState, int64, *utils.Big) {
 	return NodeStateUnreachable, -1, nil
 }
 
+func (e *erroringNode) PriorityLevel() int32 {
+	return 0
+}
+
 func (e *erroringNode) DeclareOutOfSync()            {}
 func (e *erroringNode) DeclareInSync()               {}
 func (e *erroringNode) DeclareUnreachable()          {}