+++ core/chains/evm/client/node.go
@@ -92,6 +92,7 @@ type Node interface {
 	// Name is a unique identifier for this node.
 	Name() string
 	ChainID() *big.Int
+	PriorityLevel() int32
 
 	CallContext(ctx context.Context, result interface{}, method string, args ...interface{}) error
 	BatchCallContext(ctx context.Context, b []rpc.BatchElem) error
@@ -129,12 +130,13 @@ type rawclient struct {
 // It must have a ws url and may have a http url
 type node struct {
 	utils.StartStopOnce
-	lfcLog  logger.Logger
-	rpcLog  logger.Logger
-	name    string
-	id      int32
-	chainID *big.Int
-	cfg     NodeConfig
+	lfcLog        logger.Logger
+	rpcLog        logger.Logger
+	name          string
+	id            int32
+	chainID       *big.Int
+	cfg           NodeConfig
+	priorityLevel int32
 
 	ws   rawclient
 	http *rawclient
@@ -177,13 +179,14 @@ type NodeConfig interface {
 }
 
 // NewNode returns a new *node as Node
-func NewNode(nodeCfg NodeConfig, lggr logger.Logger, wsuri url.URL, httpuri *url.URL, name string, id int32, chainID *big.Int) Node {
+func NewNode(nodeCfg NodeConfig, lggr logger.Logger, wsuri url.URL, httpuri *url.URL, name string, id int32, chainID *big.Int, priorityLevel int32) Node {
 	n := new(node)
 	n.name = name
 	n.id = id
 	n.chainID = chainID
 	n.cfg = nodeCfg
 	n.ws.uri = wsuri
+	n.priorityLevel = priorityLevel
 	if httpuri != nil {
 		n.http = &rawclient{uri: *httpuri}
 	}
@@ -194,10 +197,12 @@ func NewNode(nodeCfg NodeConfig, lggr logger.Logger, wsuri url.URL, httpuri *url
 		"nodeName", name,
 		"node", n.String(),
 		"evmChainID", chainID,
+		"priority", n.priorityLevel,
 	)
 	n.lfcLog = lggr.Named("Lifecycle")
 	n.rpcLog = lggr.Named("RPC")
 	n.stateLatestBlockNumber = -1
+
 	return n
 }
 
@@ -1104,4 +1109,8 @@ func (n *node) String() string {
 
 func (n *node) Name() string {
 	return n.name
+}
+
+func (n *node) PriorityLevel() int32 {
+	return n.priorityLevel
 }