+++ core/chains/evm/config/v2/config.go
@@ -151,6 +151,9 @@ func legacyNode(n *Node, chainID *utils.Big) (v2 types.Node) {
 	if n.SendOnly != nil {
 		v2.SendOnly = *n.SendOnly
 	}
+	if n.Priority != nil {
+		v2.PriorityLevel = *n.Priority
+	}
 	return
 }
 
@@ -719,6 +722,7 @@ type Node struct {
 	WSURL    *models.URL
 	HTTPURL  *models.URL
 	SendOnly *bool
+	Priority *int32
 }
 
 func (n *Node) ValidateConfig() (err error) {
@@ -760,6 +764,13 @@ func (n *Node) ValidateConfig() (err error) {
 		}
 	}
 
+	if n.Priority != nil && *n.Priority < 0 {
+		err = multierr.Append(err, v2.ErrInvalid{Name: "Priority", Value: *n.Priority, Msg: "priority cannot be a negative number"})
+	} else if n.Priority == nil {
+		z := int32(0)
+		n.Priority = &z
+	}
+
 	return
 }
 
@@ -775,5 +786,8 @@ func (n *Node) SetFrom(f *Node) {
 	}
 	if f.SendOnly != nil {
 		n.SendOnly = f.SendOnly
+	}
+	if f.Priority != nil {
+		n.Priority = f.Priority
 	}
 }